GitPod config with Dockerfile
Current configuration with .gitpod.yml requires the user to wait a while after gitpod workspace indicates "ready" which is not good according to @lhstrh. This change allows gitpod to not show stuff before time-consuming configurations are completed. This is also a crucial step towards building a docker image that supports LF development.

1